[tint][ir][val] Add kIsConverter check for `convert`
Adds additional restrictions when looking up match in intrinsic table
for `convert`, so only converters, not constructors, are used. This
causes identity converts, T -> T, to fail, so existing code that
generates them is updated.

Fixes: 389589696
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/222216
Auto-Submit: Ryan Harrison <rharrison@chromium.org>
Commit-Queue: Ryan Harrison <rharrison@chromium.org>

+++ src/tint/lang/hlsl/writer/access_test.cc
@@ -958,7 +958,7 @@ void foo() {
   uint x = 1u;
   vector<float16_t, 4> a = tint_bitcast_to_f16(v[0u].xy);
   float16_t b = float16_t(f16tof32(v[0u].x));
-  uint v_5 = (uint(min(x, 3u)) * 2u);
+  uint v_5 = (min(x, 3u) * 2u);
   uint v_6 = v[(v_5 / 16u)][((v_5 % 16u) / 4u)];
   float16_t c = float16_t(f16tof32((v_6 >> ((((v_5 % 4u) == 0u)) ? (0u) : (16u)))));
   float16_t d = float16_t(f16tof32(v[0u].y));
