+++ test/tint/buffer/uniform/std140/array/mat3x4_f32/dynamic_index_via_ptr.wgsl.expected.ir.dxc.hlsl
@@ -37,8 +37,8 @@ ary_ret v_1(uint start_byte_offset) {
 
 [numthreads(1, 1, 1)]
 void f() {
-  uint v_5 = (48u * uint(min(uint(i()), 3u)));
-  uint v_6 = (16u * uint(min(uint(i()), 2u)));
+  uint v_5 = (48u * min(uint(i()), 3u));
+  uint v_6 = (16u * min(uint(i()), 2u));
   float3x4 l_a[4] = v_1(0u);
   float3x4 l_a_i = v(v_5);
   float4 l_a_i_i = asfloat(a[((v_5 + v_6) / 16u)]);
