+++ test/tint/buffer/uniform/std140/struct/mat2x3_f16/dynamic_index_via_ptr.wgsl.expected.ir.dxc.hlsl
@@ -94,9 +94,9 @@ ary_ret_1 v_18(uint start_byte_offset) {
 
 [numthreads(1, 1, 1)]
 void f() {
-  uint v_23 = (256u * uint(min(uint(i()), 3u)));
-  uint v_24 = (64u * uint(min(uint(i()), 3u)));
-  uint v_25 = (8u * uint(min(uint(i()), 1u)));
+  uint v_23 = (256u * min(uint(i()), 3u));
+  uint v_24 = (64u * min(uint(i()), 3u));
+  uint v_25 = (8u * min(uint(i()), 1u));
   Outer l_a[4] = v_18(0u);
   Outer l_a_i = v_15(v_23);
   Inner l_a_i_a[4] = v_10(v_23);
@@ -104,7 +104,7 @@ void f() {
   matrix<float16_t, 2, 3> l_a_i_a_i_m = v_4((v_23 + v_24));
   uint4 v_26 = a[(((v_23 + v_24) + v_25) / 16u)];
   vector<float16_t, 3> l_a_i_a_i_m_i = tint_bitcast_to_f16((((((((v_23 + v_24) + v_25) % 16u) / 4u) == 2u)) ? (v_26.zw) : (v_26.xy))).xyz;
-  uint v_27 = (((v_23 + v_24) + v_25) + (uint(min(uint(i()), 2u)) * 2u));
+  uint v_27 = (((v_23 + v_24) + v_25) + (min(uint(i()), 2u) * 2u));
   uint v_28 = a[(v_27 / 16u)][((v_27 % 16u) / 4u)];
   float16_t l_a_i_a_i_m_i_i = float16_t(f16tof32((v_28 >> ((((v_27 % 4u) == 0u)) ? (0u) : (16u)))));
 }
