+++ test/tint/buffer/uniform/std140/struct/mat3x2_f16/dynamic_index_via_ptr.wgsl.expected.ir.dxc.hlsl
@@ -89,16 +89,16 @@ ary_ret_1 v_15(uint start_byte_offset) {
 
 [numthreads(1, 1, 1)]
 void f() {
-  uint v_20 = (256u * uint(min(uint(i()), 3u)));
-  uint v_21 = (64u * uint(min(uint(i()), 3u)));
-  uint v_22 = (4u * uint(min(uint(i()), 2u)));
+  uint v_20 = (256u * min(uint(i()), 3u));
+  uint v_21 = (64u * min(uint(i()), 3u));
+  uint v_22 = (4u * min(uint(i()), 2u));
   Outer l_a[4] = v_15(0u);
   Outer l_a_i = v_12(v_20);
   Inner l_a_i_a[4] = v_7(v_20);
   Inner l_a_i_a_i = v_5((v_20 + v_21));
   matrix<float16_t, 3, 2> l_a_i_a_i_m = v_2((v_20 + v_21));
   vector<float16_t, 2> l_a_i_a_i_m_i = tint_bitcast_to_f16(a[(((v_20 + v_21) + v_22) / 16u)][((((v_20 + v_21) + v_22) % 16u) / 4u)]);
-  uint v_23 = (((v_20 + v_21) + v_22) + (uint(min(uint(i()), 1u)) * 2u));
+  uint v_23 = (((v_20 + v_21) + v_22) + (min(uint(i()), 1u) * 2u));
   uint v_24 = a[(v_23 / 16u)][((v_23 % 16u) / 4u)];
   float16_t l_a_i_a_i_m_i_i = float16_t(f16tof32((v_24 >> ((((v_23 % 4u) == 0u)) ? (0u) : (16u)))));
 }
