+++ test/tint/buffer/uniform/std140/struct/mat4x2_f16/dynamic_index_via_ptr.wgsl.expected.ir.dxc.hlsl
@@ -90,16 +90,16 @@ ary_ret_1 v_16(uint start_byte_offset) {
 
 [numthreads(1, 1, 1)]
 void f() {
-  uint v_21 = (256u * uint(min(uint(i()), 3u)));
-  uint v_22 = (64u * uint(min(uint(i()), 3u)));
-  uint v_23 = (4u * uint(min(uint(i()), 3u)));
+  uint v_21 = (256u * min(uint(i()), 3u));
+  uint v_22 = (64u * min(uint(i()), 3u));
+  uint v_23 = (4u * min(uint(i()), 3u));
   Outer l_a[4] = v_16(0u);
   Outer l_a_i = v_13(v_21);
   Inner l_a_i_a[4] = v_8(v_21);
   Inner l_a_i_a_i = v_6((v_21 + v_22));
   matrix<float16_t, 4, 2> l_a_i_a_i_m = v_2((v_21 + v_22));
   vector<float16_t, 2> l_a_i_a_i_m_i = tint_bitcast_to_f16(a[(((v_21 + v_22) + v_23) / 16u)][((((v_21 + v_22) + v_23) % 16u) / 4u)]);
-  uint v_24 = (((v_21 + v_22) + v_23) + (uint(min(uint(i()), 1u)) * 2u));
+  uint v_24 = (((v_21 + v_22) + v_23) + (min(uint(i()), 1u) * 2u));
   uint v_25 = a[(v_24 / 16u)][((v_24 % 16u) / 4u)];
   float16_t l_a_i_a_i_m_i_i = float16_t(f16tof32((v_25 >> ((((v_24 % 4u) == 0u)) ? (0u) : (16u)))));
 }
