+++ test/tint/builtins/atomics/from_gen/var/atomicStore/storage_u32.spvasm.expected.ir.dxc.hlsl
@@ -8,7 +8,7 @@ void atomicStore_cdc29e() {
   arg_1 = 1u;
   uint x_18 = arg_1;
   uint v = 0u;
-  sb_rw.InterlockedExchange(uint(0u), x_18, v);
+  sb_rw.InterlockedExchange(0u, x_18, v);
 }
 
 void fragment_main_1() {
@@ -29,7 +29,7 @@ void atomicStore_cdc29e() {
   arg_1 = 1u;
   uint x_18 = arg_1;
   uint v = 0u;
-  sb_rw.InterlockedExchange(uint(0u), x_18, v);
+  sb_rw.InterlockedExchange(0u, x_18, v);
 }
 
 void compute_main_1() {
