+++ test/tint/expressions/bitcast/let/32bit/u32-vec2f16.wgsl.expected.glsl
@@ -2,7 +2,7 @@
 #extension GL_AMD_gpu_shader_half_float: require
 
 f16vec2 tint_bitcast_to_f16(uint src) {
-  return unpackFloat2x16(uint(src));
+  return unpackFloat2x16(src);
 }
 layout(local_size_x = 1, local_size_y = 1, local_size_z = 1) in;
 void main() {
